+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\index.js":"1","C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\resetStyle\\ResetStyles.js":"2","C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\pages\\Templates.js":"3","C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\pages\\WritingStuff.js":"4","C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\styles\\LoginStyle.js":"5","C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\profile\\ProfileModifyPage.js":"6","C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\styles\\RegisterStyle.js":"7","C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\styles\\ProfileStyle.js":"8","C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\styles\\ProfileModifyStyle.js":"9","C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\login\\LoginPage.js":"10","C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\pages\\Attention.js":"11","C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\pages\\MainList.js":"12"},{"size":758,"mtime":1606661590583,"results":"13","hashOfConfig":"14"},{"size":2545,"mtime":1606660900775,"results":"15","hashOfConfig":"14"},{"size":446,"mtime":1606661590589,"results":"16","hashOfConfig":"14"},{"size":3094,"mtime":1606661646609,"results":"17","hashOfConfig":"14"},{"size":3038,"mtime":1606661738493,"results":"18","hashOfConfig":"14"},{"size":757,"mtime":1606660900772,"results":"19","hashOfConfig":"14"},{"size":3542,"mtime":1606661807519,"results":"20","hashOfConfig":"14"},{"size":2497,"mtime":1606661815916,"results":"21","hashOfConfig":"14"},{"size":2105,"mtime":1606661824079,"results":"22","hashOfConfig":"14"},{"size":1137,"mtime":1606660900765,"results":"23","hashOfConfig":"14"},{"size":616,"mtime":1606660900768,"results":"24","hashOfConfig":"14"},{"size":324,"mtime":1606662402515,"results":"25","hashOfConfig":"14"},{"filePath":"26","messages":"27","errorCount":1,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"h48rkq",{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"34"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"43","messages":"44","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"45","messages":"46","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"47","messages":"48","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"49","messages":"50","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\index.js",["51"],"C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\resetStyle\\ResetStyles.js",[],"C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\pages\\Templates.js",[],"C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\pages\\WritingStuff.js",[],["52","53","54","55","56","57"],"C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\styles\\LoginStyle.js",[],"C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\profile\\ProfileModifyPage.js",[],"C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\styles\\RegisterStyle.js",[],"C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\styles\\ProfileStyle.js",[],"C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\styles\\ProfileModifyStyle.js",[],"C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\login\\LoginPage.js",[],"C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\pages\\Attention.js",[],"C:\\Users\\yoon_\\OneDrive\\Desktop\\WebData\\danggeun-market\\src\\pages\\MainList.js",[],{"ruleId":null,"fatal":true,"severity":2,"message":"58","line":6,"column":10},{"ruleId":"59","replacedBy":"60"},{"ruleId":"61","replacedBy":"62"},{"ruleId":"63","replacedBy":"64"},{"ruleId":"65","replacedBy":"66"},{"ruleId":"67","replacedBy":"68"},{"ruleId":"69","replacedBy":"70"},"Parsing error: Identifier 'Provider' has already been declared\n\n  4 | import { Provider } from 'react-redux';\n  5 | import { BrowserRouter } from 'react-router-dom';\n> 6 | import { Provider } from 'react-redux';\n    |          ^\n  7 | import { createStore } from 'redux';\n  8 | import { composeWithDevTools } from 'redux-devtools-extension';\n  9 | import App from './App';","lines-around-directive",["71"],"no-spaced-func",["72"],"global-require",[],"no-buffer-constructor",[],"no-new-require",[],"no-path-concat",[],"padding-line-between-statements","func-call-spacing"]
