+++ package.json
@@ -8,16 +8,16 @@
     "@testing-library/user-event": "^12.1.10",
     "babel-eslint": "^10.1.0",
     "eslint-config-airbnb": "^18.2.1",
-    "firebase": "^8.1.1",
     "eslint-config-prettier": "^6.15.0",
+    "firebase": "^8.1.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-redux": "^7.2.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
     "redux": "^4.0.5",
-    "redux-devtools": "^3.7.0",
     "redux-actions": "^2.6.5",
+    "redux-devtools": "^3.7.0",
     "redux-devtools-extension": "^2.13.8",
     "styled-components": "^5.2.1",
     "styled-reset": "^4.3.1",
