+++ src/styles/ProfileModifyStyle.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import x from '../images/ico/ic_window_close_outline_18.png';
+import x from '../images/ico/ico_window_close_outline_18.png';
 import danggeun from '../images/ico/ico_logo_72x72.png';
 import loopy from '../images/ico/loopy.jpeg';
